[2.0.3] Release

+++ modelgen/src/test/java/cz/cvut/kbss/jopa/modelgen/ModelGenProcessorTests.java
@@ -19,7 +19,7 @@ package cz.cvut.kbss.jopa.modelgen;
 
 import com.google.testing.compile.Compilation;
 import com.google.testing.compile.JavaFileObjects;
-import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
@@ -27,46 +27,48 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.List;
 
 import static com.google.testing.compile.CompilationSubject.assertThat;
 import static com.google.testing.compile.Compiler.javac;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
 class ModelGenProcessorTests {
 
-    static String actualResult = "";
+    private static final String OUTPUT_DIRECTORY = "target/generated-test-sources/static-metamodel";
 
-    @AfterAll
-    static void deleteSMClass() {
-        File file = new File("./target/generated-test-sources/static-metamodel/cz/test/ex/TestingClass_.java");
-        file.delete();
+    @AfterEach
+    void deleteSMClass() {
+        File file = new File(OUTPUT_DIRECTORY + "/cz/test/ex/TestingClassOWL_.java");
+        if (file.exists()) {
+            file.delete();
+        }
     }
 
     @Test
-    void createSMClassSuccess() throws Exception {
-        String outputDirectory = "./target/generated-test-sources/static-metamodel";
-
-        List<String> options = new ArrayList<>();
-        options.add("-AoutputDirectory=" + outputDirectory);
-        options.add("-AdebugOption=" + "true");
-        Compilation compilation =
-                javac()
-                        .withProcessors(new ModelGenProcessor()).withOptions(options)
-                        .compile(
-                                JavaFileObjects.forSourceLines("cz.test.ex.TestingClassOWL", readFileAsString(
-                                        new File("src/test/java/cz/test/ex/TestingClassOWL.java"))),
-                                JavaFileObjects.forSourceLines("cz.test.ex.TestingClassNonEntity", readFileAsString(
-                                        new File("src/test/java/cz/test/ex/TestingClassNonEntity.java"))),
-                                JavaFileObjects.forSourceLines("cz.test.ex.TestingClassNotOWL", readFileAsString(
-                                        new File("src/test/java/cz/test/ex/TestingClassNotOWL.java"))));
+    void metamodelGenerationProducesCompilableStaticMetamodelClasses() throws Exception {
+        List<String> options = List.of("-AoutputDirectory=" + OUTPUT_DIRECTORY, "-AdebugOption=" + "true");
+        Compilation compilation = javac()
+                .withProcessors(new ModelGenProcessor()).withOptions(options)
+                .compile(JavaFileObjects.forSourceLines("cz.test.ex.TestingClassOWL", readFileAsString(
+                                new File("src/test/java/cz/test/ex/TestingClassOWL.java"))),
+                        JavaFileObjects.forSourceLines("cz.test.ex.TestingClassNonEntity", readFileAsString(
+                                new File("src/test/java/cz/test/ex/TestingClassNonEntity.java"))),
+                        JavaFileObjects.forSourceLines("cz.test.ex.TestingClassNotOWL", readFileAsString(
+                                new File("src/test/java/cz/test/ex/TestingClassNotOWL.java"))));
         assertThat(compilation).succeededWithoutWarnings();
 
-        assertFalse(Files.exists(Paths.get(outputDirectory + "/cz/test/ex/TestingClassNonEntity_.java")));
-        assertFalse(Files.exists(Paths.get(outputDirectory + "/cz/test/ex/TestingClassNotOWL_.java")));
+        assertFalse(Files.exists(Paths.get(OUTPUT_DIRECTORY + "/cz/test/ex/TestingClassNonEntity_.java")));
+        assertFalse(Files.exists(Paths.get(OUTPUT_DIRECTORY + "/cz/test/ex/TestingClassNotOWL_.java")));
 
-        actualResult = readFileAsString(new File(outputDirectory + "/cz/test/ex/TestingClassOWL_.java"));
+        final Compilation staticMetamodelCompilation = javac()
+                .compile(JavaFileObjects.forSourceLines("cz.test.ex.TestingClassOWL_",
+                        readFileAsString(new File(OUTPUT_DIRECTORY + "/cz/test/ex/TestingClassOWL_.java"))));
+
+        assertThat(staticMetamodelCompilation).succeededWithoutWarnings();
+        assertEquals(readFileAsString(new File("src/test/resources/TestingClassOWL_.txt")),
+                readFileAsString(new File(OUTPUT_DIRECTORY + "/cz/test/ex/TestingClassOWL_.java")));
     }
 
     static String readFileAsString(File file) throws IOException {
